+++ dns/dnssecalgs/ecdsa.py
@@ -1,11 +1,10 @@
-from dns.dnssec.algbase import AlgorithmPrivateKeyBase, AlgorithmPublicKeyBase
-
 from dataclasses import dataclass
+
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.asymmetric import ec
-from cryptography.hazmat.primitives.asymmetric import utils
+from cryptography.hazmat.primitives.asymmetric import ec, utils
 
+from dns.dnssecalgs.base import AlgorithmPrivateKeyBase, AlgorithmPublicKeyBase
 from dns.dnssectypes import Algorithm
 from dns.rdtypes.ANY.DNSKEY import DNSKEY
