+++ dns/dnssecalgs/eddsa.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import ed448, ed25519
 
-from dns.dnssec.algbase import AlgorithmPrivateKeyBase, AlgorithmPublicKeyBase
+from dns.dnssecalgs.base import AlgorithmPrivateKeyBase, AlgorithmPublicKeyBase
 from dns.dnssectypes import Algorithm
 from dns.rdtypes.ANY.DNSKEY import DNSKEY
